Don't create org dirs by default
Require a flag to create org directories. This tools is meant to mainly
be used to update existing repositories not do a full clone of
everything from scratch. If we detect a organization directory is
missing, print an error and exit and require a '-f' to work.

Go by organization directory instead of full path since we want to clone
new repos in existing organizations.