Add an option to run a recursive clone
This commit adds an option to run git clone with the --recursive flag.
This is used when there is a git submodule in a repo and it will also
clone the submodule code in addition to the parent repo. Right now there
the submodule will just be a blank directory when the parent repo with
submodules is cloned.